More typechecker tests, support for failing tests
This CL adds support for marking tests as known failures using the following syntax:

/*expectation1::shouldbe::expectation2*/

where expectation is either "level:StaticInfo" as before, or "pass".

The meaning is that the test is currently exhibiting incorrect behavior expectation1, but should exhibit behavior expectation2.

The test will succeed so long as expectation1 is matched, but a summary of the number of "shouldbe" expectations is printed at the end of running the test script.  Currently this is just implemented as grep, we can make this fancier eventually.

I'm not particularly attached to the syntax, feel free to suggest alternatives.

This CL also adds a bunch of fairly mechanical tests of subtyping, a number of which we are definitely failing, and some of which we may or may not be failing depending on where we wish to insert automatic fixup code.

BUG=
R=vsm@google.com

Review URL: https://chromereviews.googleplex.com/129027013