Closure wrapping, revamped coercions
This CL adds composite coercions which reify the choices made by the typechecker in allowing a closure wrapping as explicit casts and wrappers.  The code generators should now be able to just walk the coercion generating the appropriate code.  Thoughts or suggestions about the design of this very welcome.

The closure wrapping decision process has changed as well.  Instead of simply checking reverse subtyping (which allowed some closure wrapping which would change the runtime type), we now essentially do a more extended subtype check which allows downcasts (and a few other conversions) on arguments, and attempts to wrap or downcast the body.  This enables wrapping for a common pattern we have encountered in which an untyped (and hence dynamically typed) lambda is passed into a typed context.  It also allows closure wrapping for some cases that we may (or may not) prefer to reject, such as allowing an int -> int to be coerced to an Object -> Object and vice versa.

In follow up CLs, I will add a reifier which translates coercions to ast nodes, and will add some flags to try out different closure wrapping strategies.

BUG=
R=vsm@google.com

Review URL: https://chromereviews.googleplex.com/132147015