Coercion reification, dart codegen URI rewriting, import fixups.
This CL adds a general reification pass which makes abstract coercions concrete by modifiying the ast.  Casts are reified as "as" expressions, and wrappers are reified by creating functions which take the old function as arguments and produce a new wrapped function.  There are a number of known issues not yet addressed:

1) No distinction between dart "as" and our special "as" is made yet.
2) Wrapper generation functions do not check for null arguments yet.
3) Types introduced via wrappers and "as" expressions are not yet added to the set of types for which new imports may be required.
4) Wrapper functions for non-literals can't be guaranteed not to change the runtime type without more backend support, since we only have their static type.

This CL also adds support for rewriting import directives in the dart backend to reflect the directory structure that we currently use for output.  Probably this should be rethought.  Currently I just use absolute file URIs, which is not what we want.

This CL also adds support for adding in the import directives needed to account for type names added via inference.  This is not yet correct for "part of" files, since we need to add the imports to the main library file.

BUG=
R=vsm@google.com

Review URL: https://chromereviews.googleplex.com/137957013